requirements_dev.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.20
django-cors-headers==2.5.0
django-extensions==2.1.6
django-filter==2.1.0
django-nose==1.4.6
django-rest-swagger==2.2.0
djangorestframework==3.9.2
elasticsearch==6.3.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mock==2.0.0
nose==1.3.7
openapi-codec==1.3.2
pbr==5.1.3
pytz==2018.9
requests==2.21.0
restless==2.1.1
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
